refactor: gather templates
- UI 관련 템플릿 모아서 작성
